% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpsReport.r
\name{carpsReport}
\alias{carpsReport}
\title{carpsReport function}
\usage{
carpsReport(Report_Type, Article_ID, Insufficient_Information_Errors,
  Decision_Errors, Major_Numerical_Errors, Time_to_Complete,
  Substantive_Conclusion_Affected, Author_Assistance)
}
\arguments{
\item{Report_Type}{Enter 'pilot' or 'final'}

\item{Article_ID}{Enter the article's unique ID code}

\item{Insufficient_Information_Errors}{Enter the number of Insufficient Information Errors}

\item{Decision_Errors}{Enter the number of decision errors}

\item{Major_Numerical_Errors}{Enter the number of major numerical errors}

\item{Time_to_Complete}{Enter the estimated time to complete the report in minutes}

\item{Substantive_Conclusion_Affected}{Enter whether you think the substantive conclusion of the article is affected by any errors you encountered (T/F)}

\item{Author_Assistance}{Enter whether author assistance was required (T/F)}
}
\value{
Returns a formatted table (via kable) reporting error tallys. Also writes out a csv object containing the error tallys.
}
\description{
This is a helper function for building standardised final outcomes for CARPS reproducibility reports.
}
\examples{
carpsReport(Report_Type = "pilot", Article_ID = "ABhgyo", Insufficient_Information_Errors = 0, Decision_Errors = 1, Major_Numerical_Errors = 4, Time_to_Complete = 120, Substantive_Conclusion_Affected = F, Author_Assistance = T)
}
